Add interface to the new service and remove form db repository
