+++ backend/dataall/modules/datasets/services/dataset_service.py
@@ -1,7 +1,8 @@
 import os
 import json
 import logging
-
+from typing import List
+from abc import ABC, abstractmethod
 from dataall.base.aws.quicksight import QuicksightClient
 from dataall.base.db import exceptions
 from dataall.base.utils.naming_convention import NamingConventionPattern
@@ -48,8 +49,37 @@ from dataall.modules.datasets.services.dataset_permissions import DATASET_TABLE_
 
 log = logging.getLogger(__name__)
 
+class DatasetServiceInterface(ABC):
+    @staticmethod
+    def check_before_delete(self, session, uri, **kwargs):
+        """Abstract method to be implemented by dependent modules that want to add checks before deletion for dataset objects"""
+        return True
+
+    @staticmethod
+    def execute_on_delete(self, session, uri, **kwargs):
+        """Abstract method to be implemented by dependent modules that want to add clean-up actions when a dataset object is deleted"""
+        return True
 
 class DatasetService:
+    _interfaces: List[DatasetServiceInterface] = []
+
+    @classmethod
+    def register(cls, interface: DatasetServiceInterface):
+        cls._interfaces.append(interface)
+
+    @classmethod
+    def check_before_delete(cls, session, uri, **kwargs) -> bool:
+        """All actions form other modules that need to be executed before deletion"""
+        can_be_deleted = [interface.check_before_delete(session, uri, **kwargs) for interface in cls._interfaces]
+        return False not in set(can_be_deleted)
+
+    @classmethod
+    def execute_on_delete(cls, session, uri, **kwargs) -> bool:
+        """All actions form other modules that need to be executed during deletion"""
+        for interface in cls._interfaces:
+            interface.execute_on_delete(session, uri, **kwargs)
+        return True
+
     @staticmethod
     def check_dataset_account(session, environment):
         dashboards_enabled = EnvironmentService.get_boolean_env_param(session, environment, 'dashboardsEnabled')
@@ -376,7 +406,7 @@ class DatasetService:
         with context.db_engine.scoped_session() as session:
             dataset: Dataset = DatasetRepository.get_dataset_by_uri(session, uri)
             env = EnvironmentService.get_environment_by_uri(session, dataset.environmentUri)
-            DatasetRepository.check_before_delete(session, uri, action=DELETE_DATASET)
+            DatasetService.check_before_delete(session, uri, action=DELETE_DATASET)
 
             tables = [t.tableUri for t in DatasetRepository.get_dataset_tables(session, uri)]
             for tableUri in tables:
@@ -388,7 +418,7 @@ class DatasetService:
 
             DatasetIndexer.delete_doc(doc_id=uri)
 
-            DatasetRepository.execute_on_delete(session, uri, action=DELETE_DATASET)
+            DatasetService.execute_on_delete(session, uri, action=DELETE_DATASET)
             DatasetService.delete_dataset_term_links(session, uri)
             DatasetTableRepository.delete_dataset_tables(session, dataset.datasetUri)
             DatasetLocationRepository.delete_dataset_locations(session, dataset.datasetUri)
